removed files

+++ pure_mp3/src/pure_mp3/MusicPlayer.java
@@ -32,7 +32,7 @@ public interface MusicPlayer
 	public boolean isStopped();
 	public void setVolume(float xVolume);
 	public AudioFormat getAudioFormat();
-	public int getFramePosition();
+	public long getFramePosition();
 	public long getFrameLength();
 	public int getDurationInSeconds();
 	public Song getCurrentSong();
